+++ tester/liblinphone_tester.h
@@ -245,6 +245,7 @@ typedef struct _stats {
 	int number_of_LinphoneIsComposingActiveReceived;
 	int number_of_LinphoneIsComposingIdleReceived;
 	int progress_of_LinphoneFileTransfer;
+	int number_of_LinphoneFileTransfer;
 
 	int number_of_LinphoneChatRoomConferenceJoined;
 	int number_of_LinphoneChatRoomEphemeralTimerStarted;